Add durations to packet_in and out tests and other fixes
* Adds a duration argument to the packet_in and packet_out tests, test
duration is now linked to when generation starts not when the
openflow session is established
* Adds buffered packet-in support to the packet-in test
* Minor fixes for packet generation, in particular the pktgen_hdr size.
* Adds a barrier req/reply to packet_in test and waits until received
(or a timeout) before starting generation
* Removes unused arguments
* Fixes argument parsing bugs in packet_out
* Some other minor bug/consistancy fixes

+++ control.cpp
@@ -67,7 +67,7 @@ public:
 static int check_of_version_allowed(oflops_context *ctx, uint8_t v) {
     if (ctx->nb_of_versions == 0)
         return 1;
-    for (int i = 0; i < ctx->nb_of_versions; i++) {
+    for (size_t i = 0; i < ctx->nb_of_versions; i++) {
         if (v == ctx->of_versions[i])
             return 1;
     }
